+++ services/orchest-webserver/client/src/contexts/__tests__/ProjectsContext.test.tsx
@@ -1,19 +1,15 @@
-import { act, renderHook } from "@testing-library/react-hooks";
-import * as React from "react";
-import { PipelineMetaData } from "../../types";
+import { PipelineMetaData } from "@/types";
 import {
   chance,
   getPipelineMedadatas,
   mockProjects,
-} from "../../__mocks__/mockProjects.mock";
+} from "@/__mocks__/mockProjects.mock";
+import { act, renderHook } from "@testing-library/react-hooks";
+import * as React from "react";
 import {
   ProjectsContextProvider,
   useProjectsContext,
 } from "../ProjectsContext";
-
-const wrapper = ({ children }) => {
-  return <ProjectsContextProvider>{children}</ProjectsContextProvider>;
-};
 
 let mockData: {
   project1Uuid: string;
@@ -50,6 +46,9 @@ const resetMock = () => {
 };
 
 describe("useProjectsContext", () => {
+  const wrapper = ({ children }) => {
+    return <ProjectsContextProvider>{children}</ProjectsContextProvider>;
+  };
   const { result, waitForNextUpdate, rerender, unmount } = renderHook(
     () => useProjectsContext(),
     { wrapper }
