+++ services/orchest-webserver/client/src/project-selector/useFetchProjectsForSelector.tsx
@@ -1,5 +1,5 @@
 import { useProjectsContext } from "@/contexts/ProjectsContext";
-import { useFetchProjects } from "@/projects-view/hooks/useFetchProjects";
+import { useFetchProjects } from "@/hooks/useFetchProjects";
 import React from "react";
 
 export const useFetchProjectsForSelector = (shouldFetch: boolean) => {
